use centralized error management
