--- projects/cas3-core/lang/aliases.sexp
@@ -1,4 +1,0 @@
-(set ra replace_all)
-(set rr replace_repeated)
-(set list List)
-(set alts Alternatives)
