--- projects/cas3-core/lang/attrs.sexp
@@ -1,11 +1,0 @@
-(set (attrs setd) (list HoldAll SequenceHold))
-(set (attrs clear) (list HoldAll))
-(set (attrs hold) (list HoldAll))
-(set (attrs pattern) (list HoldFirst))
-
-(set (attrs true) (list locked protected))
-(set (attrs false) (list locked protected))
-
-(set (attrs rule) (list SequenceHold))
-(set (attrs Table) (list HoldAll))
-(set (attrs Timing) (list HoldAll))
