+++ projects/cas3-jupyter/src/executor/mod.rs
@@ -1,6 +1,6 @@
 use crate::{config::Cas3Config};
 use jupyter::{to_value, value_type::HtmlText, ExecutionRequest, JupyterError, JupyterKernelSockets, JupyterMessage};
-use cas3_core::Cas3VM;
+use cas3_core::{Cas3VM, Expr};
 
 pub struct CasExecutor {
     pub(crate) vm: Cas3VM,
@@ -26,48 +26,15 @@ impl CasExecutor {
         Ok(())
     }
 
-    pub(crate) async fn send_value(&self, value: ValkyrieValue, parent: &JupyterMessage) {
+    pub(crate) async fn send_value(&self, value: Expr, parent: &JupyterMessage) {
         match value {
-            // never type never sends
-            ValkyrieValue::Nothing => {}
-            ValkyrieValue::Null => self.sockets.send_executed(DisplayKeywords::new("null"), parent).await,
-            ValkyrieValue::Unit => self.sockets.send_executed(DisplayKeywords::new("( )"), parent).await,
-            ValkyrieValue::Boolean(v) => self.sockets.send_executed(DisplayKeywords::new(v), parent).await,
-            ValkyrieValue::Number(v) => self.sockets.send_executed(DisplayNumber::new(v), parent).await,
-            ValkyrieValue::Unicode(v) => self.sockets.send_executed(v.to_string(), parent).await,
-            ValkyrieValue::UTF8String(v) => self.sockets.send_executed(v.get().clone(), parent).await,
-            ValkyrieValue::Bytes(_) => {
-                todo!()
+            Expr::Int(v) => {
+                self.sockets.send_executed(v.to_string(), parent).await
             }
-            ValkyrieValue::Class(_) => {
-                todo!()
-            }
-            ValkyrieValue::Variant(_) => {
-                todo!()
-            }
-            ValkyrieValue::NDArray(_) => {
-                todo!()
-            }
-            ValkyrieValue::Image(_) => {
-                todo!()
-            }
-            ValkyrieValue::Html(v) => {
-                self.sockets.send_executed(HtmlText::new(v), parent).await;
-            }
-            ValkyrieValue::Uninitialized => {
-                todo!()
-            }
-            ValkyrieValue::List(v) => match to_value(v) {
-                Ok(o) => self.sockets.send_executed(o, parent).await,
-                Err(_) => {}
-            },
-            ValkyrieValue::Dict(v) => match to_value(v) {
-                Ok(o) => self.sockets.send_executed(o, parent).await,
-                Err(_) => {}
-            },
-            // ValkyrieValue::DataFrame(_) => {
-            //     todo!()
-            // }
+            Expr::Real(v) => { self.sockets.send_executed(v.to_string(), parent).await }
+            Expr::Sym(v) => { self.sockets.send_executed(v.to_string(), parent).await }
+            Expr::Str(v) => { self.sockets.send_executed(v.to_string(), parent).await }
+            Expr::List(v) => { self.sockets.send_executed(v.to_string(), parent).await }
         }
     }
 }
